feat: 단일 로또 티켓(6개의 숫자)를 관리하는 도메인 구현
