+++ src/main/java/lotto/domain/model/Lotto.java
@@ -1,0 +1,39 @@
+package lotto.domain.model;
+
+import lotto.util.ErrorMessages;
+
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+
+public class Lotto {
+    private final List<Integer> numbers;
+
+    public Lotto(List<Integer> numbers) {
+        validate(numbers);
+        validateNumberRange(numbers);
+        validateDuplication(numbers);
+        this.numbers = numbers;
+    }
+
+    private void validate(List<Integer> numbers) {
+        if (numbers.size() != 6) {
+            throw new IllegalArgumentException(ErrorMessages.INVALID_WINNING_NUMBER_COUNT);
+        }
+    }
+    private void validateNumberRange(List<Integer> numbers) {
+        if (numbers.stream().anyMatch(num -> num < 1 || num > 45)) {
+            throw new IllegalArgumentException(ErrorMessages.INVALID_WINNING_NUMBER_RANGE);
+        }
+    }
+
+    private void validateDuplication(List<Integer> numbers) {
+        Set<Integer> uniqueNumbers = new HashSet<>(numbers);
+        if (uniqueNumbers.size() != numbers.size()) {
+            throw new IllegalArgumentException(ErrorMessages.DUPLICATE_WINNING_NUMBER);
+        }
+    }
+    public List<Integer> getNumbers() {
+        return numbers;
+    }
+}
