+++ src/test/java/lotto/LottoTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.model.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
